Crate rs_car_ipfs

Wrapper for rs-car to read files from IPFS trustless gateways with an async API.

Usage

To read a single file buffering the block dag single_file::read_single_file_buffer

To read a single file without buffering the block dag single_file::read_single_file_seek

Modules

single_file

CAR stream does not include duplicated blocks, so to reconstruct a unixfs file, data does not follow the same layout as the expected target file. To recreate the file one must have the ability to read from arbitrary locations of the stream.

Type Aliases

Cid

A Cid that contains a multihash with an allocated size of 512 bits.